restructure homepage components - hoisting state to move results chart
